Stray, empty tiles removed from maps
A couple of the maps had thousands of these... The initiative that
has begun to give empty/free tiles some graphics is a needed one.

A couple of "monster traps" were removed as well, I'll submit a
patch with the walkmaps later. Maps: 003-1, 013-1 and 018-1.

The three swamp/cemetery maps have not yet been checked.