correction of the automapping files.
location changed from maps/rules/* to rules/*
This was done to avoid trouble with the serverscripts of tmw.
(So there are only those maps in the maps folder which are needed for the game.)

The automapping feature got a more powerful logic.
So there were slight changes at the rules as well

In the rules folder there are 3 examples for automapping.
There is just the set layer done. After automapping the rest
of the maps is set up automatically.