Full map review:
Renamed layer names in order to have the same layer naming scheme on all maps.
In 012-3 map, removed an unnecessary layer and transported its tiles to another layer. In game, it looks the same.
All monster spawns in 025-1 and 026-1 that covered the whole map were given 0,0 coordinates.
Improving the warps position on the maps.

Started reviewing all the objects of all maps:
- Giving natural numbers to coordinates, height and width.
- Renaming correctly each object name.
- Avoid the use of capital letters in object properties and type.
- Removed collision tiles put just in front of warps.
- Changed few last map names to fit correctly with the warp names.
- Random layer fixes.