Fixed status effect icons Y offsets for the 0.6 client.
The new client is handling sprites in a unified way regarding
the Y offset. Thus, icons, which are sprites, had to be adapted.
We decided to simplify the offset in the new client so that
the handling there is more simple in the long term.

The icon will have an offset in the 0.5.x clients but it should
still be bearable.