Added attributes.xml
Used for telling the client about the minimum and maximum amount of
points each attribute can have as well as the amount of points that need
to be distributed.

The listed attributes are only used to validate the correctness of the
character starting points. We could make the client use them, but that
would be at the cost of internationalization.

See http://bugs.manasource.org/view.php?id=501