019-3 & 030-3:
	* fixing https://forums.themanaworld.org/viewtopic.php?p=140509#p140509
	* fixing https://github.com/themanaworld/tmwa-client-data/issues/50
In commit e6fe61806db6bdd8a95c64d96b12b8ba42de0d96 a wrong version of map 019-3 was used and therefore caused problems with warps to this map. I looked at the new map and tried to find changes from the commit and applied everything I found to the old version of the map. Following I had to edit the warp on 030-3.