Dark Talisman FX effect - by Hello=) of TMW.
Adds fancy rotating ring and yellow-redish glow for
players wearing Dark Talisman, so it would also have
some "aura" comparable to few other "powerful" items.

Additionally, it also fixes Dark Talisman description.
Credits for this are going to Ledmitz who both hilighted
problem, got me curious enough to fiddle with particle
engine and showcased description fix.

This commit is self-sufficient. Effect only relies on
pre-existing images that are already in the tree.

It should not cause any conflicts or problems as it
touches bare minimum of what it have to touch.