:)
We talked a bit about Amulet of Yendor in -irc and agreed that it'd be a
cool joke. But we did not agree on whether to rename the GM Amulet into
Amulet of Yendor or Rodney. Maybe, since the Yendor naming is used for
the one the player is supposed to fetch in Nethack, the GM one could be
named Rodney and thus the Yendor name remains for possible future
Amulet, should TMW gain a dungeon crawling quest. Until then, Amulet of
Rodney shall serve as inspiration.