Poppys maps and some others
includes maps poppy made for some xmas event and some dev meeting place, a jail and a map for the guildbot, not clue yet where the prison can be entered and if event maps should get their backdrop removed and filled with normal tiles but at least everything should be accessible atm so devs/gms can take a stroll. @Ledmitz dont looked if the music those maps are using is included or if it fits.
<br>
<br>
recompressed all pictures to lower size