This is client data part of Arrows Bundle item/technology.
It been talked many times on IRC, forums and so on as idea but haven't gone far due to
lack of item graphics and lack of my skills around items. However, yesterday Wellvin
authored arrows bundle graphics that is enough for me - and I did the rest.

This is also dubs as preparatory step to make swamp more appealing for archers.
(Enchanters are meant to drop this item, its subject for other MRs)

How this works:
* There's now item called "Arrows Bundle" that meant to be dropped some mobs.
* Upon use it "unbundled", bundle item removed and random # of random arrows type given.
  (so its a bit of gambling what particular bundle will be)

NB:
* Also see serverdata!907 to actually make it in effect.
* I tested whole thing on local and it works as expected.
* At some point later item should be added as drop to at least Enchanters.

Co-authored-by: Wellvin
Tested-by: Hello=)

Approved-by: Thorbjørn Lindeijer <bjorn@lindeijer.nl>